* commit '92a66cd28466e30a22a463f5c903c3625179b733':
  Changed distances that were int's to use doubles
  Switched from using a map-based data structure to associate points to verticies, to using an rtree-based data structure

+++ include/stats.h
@@ -98,6 +98,6 @@ class STATS
 	int CheckGaussians(double angle);
 	void ScanLine(VECTOR lines, int nb_scanlines);
 	void KMCluster(bool output, VECTOR lines);
-	void RasterStatistics(VECTOR lines, int dist, std::string raster_filename);
+	void RasterStatistics(VECTOR lines, double dist, std::string raster_filename);
 };
 #endif
